pyproject.toml
[tool.poetry]
name = "license_control"
version = "0.1.0"
description = ""
authors = ["KivenC <18721218083@126.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyQt5 = "^5.15.4"
pycrypto = "^2.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pyinstaller = "^4.3"
mypy = "^0.812"

[tool.pytest.ini_options]
filterwarnings = []

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
